+++ test/indeed.js
@@ -3,7 +3,7 @@ var IndeedService = require('../indeed-service.js')();
 IndeedService.query('Javascript', 'Toronto', '100')
 .then(function(data) {
     // Do something with data ...
-    console.log(data.titleList);
+    console.log(Object.keys(data));
 
 })
 .catch(function(err) {